+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Now checking if the element is scrollable before choosing default root element. [Trello 2198](https://trello.com/c/DTvpdAj4), [Trello 2207](https://trello.com/c/v5s4lv8u), [Trello 2215](https://trello.com/c/nUzTl0KB)
 - CSS stitching is now working when checking a frame overlapping with elements from its outer frame. [Trello 1846](https://trello.com/c/grlCdwMs)
 - Calling close without open now behaves properly. [Trello 2241](https://trello.com/c/h7tW49Nz)
+- Calling session delete session will work without setting a new server connector. [Trello 2246](https://trello.com/c/aANiFwRX)
 
 ## [3.181.0] - 2020-10-16
 ### Added